Removing the "and" controllflow statement from load_view
`Connection.execute` returns nil on successful execution and thus creates an infinite loop by failing `view.load!` and not executing `views.delete(view)`.

Logic is kept intact as the SQL request would exit the scope with an exception if the statement fails and so would not delete the view from the list. This way the dependency resolution code would still work fine.